Refactor : structure
